KA-100 Rename containers and remove unused code
